add dev mode for @frinx/shared package
