+++ package.json
@@ -17,12 +17,13 @@
     "dedubcheck": "dedubcheck",
     "build:api": "cd packages/frinx-api && yarn rollup -c",
     "build:shared": "cd packages/shared && yarn run build",
+    "dev:shared": "cd packages/shared && yarn run dev",
     "build:dashboard": "cd packages/frinx-dashboard && yarn run build",
     "build:server": "cd packages/frinx-frontend-server && yarn run build",
     "build": "run-s build:shared build:dashboard build:server",
     "dev:client": "cd packages/frinx-dashboard && yarn run dev",
     "dev:server": "cd packages/frinx-frontend-server && yarn run dev",
-    "dev": "run-p dev:client dev:server",
+    "dev": "run-p build:api dev:shared dev:client dev:server",
     "turbo": "turbo run typegen type-check lint",
     "pr": "run-s formatter:check version:check turbo",
     "version:check": "node ./version-check.mjs"
